+++ clients/source/v1/source.go
@@ -44,26 +44,26 @@ type FetchResultMessage struct {
 
 type ClientOption func(*Client)
 
-func WithSourceLogger(logger zerolog.Logger) func(*Client) {
+func WithLogger(logger zerolog.Logger) func(*Client) {
 	return func(c *Client) {
 		c.logger = logger
 	}
 }
 
-func WithSourceDirectory(directory string) func(*Client) {
+func WithDirectory(directory string) func(*Client) {
 	return func(c *Client) {
 		c.directory = directory
 	}
 }
 
-func WithSourceGRPCConnection(userConn *grpc.ClientConn) func(*Client) {
+func WithGRPCConnection(userConn *grpc.ClientConn) func(*Client) {
 	return func(c *Client) {
 		// we use a different variable here because we don't want to close a connection that wasn't created by us.
 		c.userConn = userConn
 	}
 }
 
-func WithSourceNoSentry() func(*Client) {
+func WithNoSentry() func(*Client) {
 	return func(c *Client) {
 		c.noSentry = true
 	}
@@ -226,12 +226,12 @@ func (c *Client) GetTables(ctx context.Context) ([]*schema.Table, error) {
 	return tables, nil
 }
 
-func (c *Client) GetDynamicTables(ctx context.Context) ([]*schema.Table, error) {
+func (c *Client) GetDynamicTables(ctx context.Context) (schema.Tables, error) {
 	res, err := c.pbClient.GetDynamicTables(ctx, &pb.GetDynamicTables_Request{})
 	if err != nil {
 		return nil, fmt.Errorf("failed to call GetDynamicTables: %w", err)
 	}
-	var tables []*schema.Table
+	var tables schema.Tables
 	if err := json.Unmarshal(res.Tables, &tables); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal tables: %w", err)
 	}